Complete all CRUD Request, handle all errors, Add Swagger for doc

+++ item-service/src/main/java/com/capstonebackend/itemservice/service/ItemService.java
@@ -1,32 +1,34 @@
 package com.capstonebackend.itemservice.service;
 
-import com.capstonebackend.itemservice.dto.ItemDTO;
 import com.capstonebackend.itemservice.entity.ItemEntity;
 import com.capstonebackend.itemservice.exception.NotFoundException;
 import com.capstonebackend.itemservice.repository.ItemRepository;
-import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.validation.Errors;
 
 import java.util.List;
 import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-@Data
 public class ItemService {
 
-    private ItemRepository itemRepository;
+    private final ItemRepository itemRepository;
 
     public List<ItemEntity> getAllItems() {
-        return itemRepository.findAll();
+        List<ItemEntity> items = itemRepository.findAll();
+        if (items != null) {
+            return items;
+        } else {
+            throw new NotFoundException(404, "Item not found");
+        }
+
     }
 
     public Optional<ItemEntity> getItemById(Long id) {
         Optional<ItemEntity> item = itemRepository.findById(id);
         if(item.isEmpty()) {
-            throw new NotFoundException("Item not found");
+            throw new NotFoundException(404, "Item not found");
         }
         return item;
     }
@@ -39,21 +41,28 @@ public class ItemService {
         Optional<ItemEntity> item = itemRepository.findItemByUuid(uuid);
 
         if (!item.isPresent()) {
-            throw new NotFoundException("Item not found");
+            throw new NotFoundException(404, "Item not found");
         }
         return item;
     }
 
     public Optional<ItemEntity> getItemByName(String name) {
-        Optional<ItemEntity> item = itemRepository.findByItemName(name);
+        Optional<ItemEntity> item = itemRepository.findItemByName(name);
 
         if (!item.isPresent()) {
-            throw new NotFoundException("Item not found");
+            throw new NotFoundException(404, "Item not found");
         }
         return item;
     }
 
-    public void deleteItem(Long id) {
+    public void deleteItemById(Long id) {
+        Optional<ItemEntity> item = itemRepository.findById(id);
+
+        if(item.isEmpty()) {
+            throw new NotFoundException(404,"Item not found");
+        }
+
         itemRepository.deleteById(id);
     }
+
 }
