+++ item-service/src/main/java/com/capstonebackend/itemservice/service/ServiceItemService.java
@@ -1,5 +1,6 @@
 package com.capstonebackend.itemservice.service;
 
+import com.capstonebackend.itemservice.entity.ItemEntity;
 import com.capstonebackend.itemservice.entity.ServiceEntity;
 import com.capstonebackend.itemservice.exception.NotFoundException;
 import com.capstonebackend.itemservice.repository.ServiceRepository;
@@ -14,16 +15,21 @@ import java.util.Optional;
 @RequiredArgsConstructor
 @Data
 public class ServiceItemService{
-    private ServiceRepository serviceRepository;
+    private final ServiceRepository serviceRepository;
 
     public List<ServiceEntity> getAllServices() {
-        return serviceRepository.findAll();
+        List<ServiceEntity> services = serviceRepository.findAll();
+        if (services != null) {
+            return services;
+        } else {
+            throw new NotFoundException(404, "service not found");
+        }
     }
 
-    public Optional<ServiceEntity> getServiceEntityById(Long id) {
+    public Optional<ServiceEntity> getServiceById(Long id) {
         Optional<ServiceEntity> service = serviceRepository.findById(id);
         if(service.isEmpty()) {
-            throw new NotFoundException("service not found");
+            throw new NotFoundException(404, "service not found");
         }
         return service;
     }
@@ -36,21 +42,27 @@ public class ServiceItemService{
         Optional<ServiceEntity> service = serviceRepository.findServiceByUuid(uuid);
 
         if (!service.isPresent()) {
-            throw new NotFoundException("service not found");
+            throw new NotFoundException(404, "service not found");
         }
         return service;
     }
 
     public Optional<ServiceEntity> getServiceByName(String name) {
-        Optional<ServiceEntity> service = serviceRepository.findByServiceName(name);
+        Optional<ServiceEntity> service = serviceRepository.findServiceByName(name);
 
         if (!service.isPresent()) {
-            throw new NotFoundException("service not found");
+            throw new NotFoundException(404, "service not found");
         }
         return service;
     }
 
-    public void deleteService(Long id) {
+    public void deleteServiceById(Long id) {
+        Optional<ServiceEntity> service = serviceRepository.findById(id);
+
+        if(service.isEmpty()) {
+            throw new NotFoundException(404,"service not found");
+        }
+
         serviceRepository.deleteById(id);
     }
 }
